+++ test/runtests.jl
@@ -10,9 +10,9 @@ x8 = randn(Float32, (256,256,8,1))
 
 @info "TinyMachines.jl tests..."
 @testset "TinyMachines.jl" begin
-    # include("./misc-tests.jl")
+    include("./misc-tests.jl")
 
-    # include("./activationlayers-tests.jl")
+    include("./activationlayers-tests.jl")
     include("./convlayers-tests.jl")
     include("./espblocks-tests.jl")
     ### TODO: irblocks-tests